FORMS-150 : Manage RBAC on all the Form Actions. (Delete, modify steps, modify config, view responses, manage publications). Add an RBAC for form creation

+++ src/java/fr/paris/lutece/plugins/forms/web/admin/FormJspBean.java
@@ -46,6 +46,7 @@ import fr.paris.lutece.plugins.forms.business.FormMessage;
 import fr.paris.lutece.plugins.forms.business.FormMessageHome;
 import fr.paris.lutece.portal.business.rbac.RBAC;
 import fr.paris.lutece.portal.business.user.AdminUser;
+import fr.paris.lutece.portal.service.admin.AccessDeniedException;
 import fr.paris.lutece.portal.service.i18n.I18nService;
 import fr.paris.lutece.portal.service.message.AdminMessage;
 import fr.paris.lutece.portal.service.message.AdminMessageService;
@@ -217,10 +218,14 @@ public class FormJspBean extends AbstractJspBean
      * @param request
      *            The Http request
      * @return the html code of the form form
+     * @throws fr.paris.lutece.portal.service.admin.AccessDeniedException
+     *            the possible AccessDeniedException
      */
     @View( VIEW_CREATE_FORM )
-    public String getCreateForm( HttpServletRequest request )
+    public String getCreateForm( HttpServletRequest request ) throws AccessDeniedException
     {
+        checkUserPermission( Form.RESOURCE_TYPE, RBAC.WILDCARD_RESOURCES_ID, FormsResourceIdService.PERMISSION_CREATE, request );
+
         _form = ( _form != null ) ? _form : new Form( );
         _formMessage = ( _formMessage != null ) ? _formMessage : new FormMessage( );
 
@@ -278,9 +283,11 @@ public class FormJspBean extends AbstractJspBean
      * @param request
      *            The Http request
      * @return the html code to confirm
+     * @throws fr.paris.lutece.portal.service.admin.AccessDeniedException
+     *                      Access denied exception if user isn't authorized
      */
     @View( VIEW_CONFIRM_REMOVE_FORM )
-    public String getConfirmRemoveForm( HttpServletRequest request )
+    public String getConfirmRemoveForm( HttpServletRequest request ) throws AccessDeniedException
     {
         int nId = -1;
         try
@@ -293,6 +300,8 @@ public class FormJspBean extends AbstractJspBean
 
             return redirectView( request, VIEW_MANAGE_FORMS );
         }
+        
+        checkUserPermission( Form.RESOURCE_TYPE, String.valueOf( nId ), FormsResourceIdService.PERMISSION_MODIFY, request);
 
         Form formToBeDeleted = FormHome.findByPrimaryKey( nId );
         String strConfirmRemoveMessage = formToBeDeleted.isActive( ) ? MESSAGE_CONFIRM_REMOVE_ACTIVE_FORM : MESSAGE_CONFIRM_REMOVE_FORM;
@@ -362,9 +371,11 @@ public class FormJspBean extends AbstractJspBean
      * @param request
      *            The Http request
      * @return the jsp URL to display the form to manage forms
+     * @throws fr.paris.lutece.portal.service.admin.AccessDeniedException
+     *           AccessDeniedException if user isn't authorized to delete the form
      */
     @Action( ACTION_REMOVE_FORM )
-    public String doRemoveForm( HttpServletRequest request )
+    public String doRemoveForm( HttpServletRequest request ) throws AccessDeniedException
     {
         int nId = -1;
         try
@@ -377,6 +388,8 @@ public class FormJspBean extends AbstractJspBean
 
             return redirectView( request, VIEW_MANAGE_FORMS );
         }
+        
+        checkUserPermission( Form.RESOURCE_TYPE, String.valueOf( nId ), FormsResourceIdService.PERMISSION_MODIFY, request);
 
         FormService formService = SpringContextService.getBean( FormService.BEAN_NAME );
         formService.removeForm( nId, getUser( ) );
@@ -392,9 +405,11 @@ public class FormJspBean extends AbstractJspBean
      * @param request
      *            The Http request
      * @return The HTML form to update info
+     * @throws fr.paris.lutece.portal.service.admin.AccessDeniedException
+     *           Access denied exception if the user isn't authorized to view the form config page
      */
     @View( VIEW_MODIFY_FORM )
-    public String getModifyForm( HttpServletRequest request )
+    public String getModifyForm( HttpServletRequest request ) throws AccessDeniedException
     {
         int nId = NumberUtils.toInt( request.getParameter( FormsConstants.PARAMETER_ID_FORM ), FormsConstants.DEFAULT_ID_VALUE );
 
@@ -402,6 +417,8 @@ public class FormJspBean extends AbstractJspBean
         {
             return redirectView( request, VIEW_MANAGE_FORMS );
         }
+        
+        checkUserPermission( Form.RESOURCE_TYPE, String.valueOf( nId ), FormsResourceIdService.PERMISSION_MODIFY_PARAMS, request);
 
         Form formToBeModified = FormHome.findByPrimaryKey( nId );
 
@@ -454,9 +471,11 @@ public class FormJspBean extends AbstractJspBean
      * @param request
      *            The Http request
      * @return The HTML form to update info
+     * @throws fr.paris.lutece.portal.service.admin.AccessDeniedException
+     *           Access denied exception if the user isn't authorized to view the form config page
      */
     @View( VIEW_MODIFY_PUBLICATION )
-    public String getModifyFormPublication( HttpServletRequest request )
+    public String getModifyFormPublication( HttpServletRequest request ) throws AccessDeniedException
     {
         int nId = NumberUtils.toInt( request.getParameter( FormsConstants.PARAMETER_ID_FORM ), FormsConstants.DEFAULT_ID_VALUE );
 
@@ -464,6 +483,8 @@ public class FormJspBean extends AbstractJspBean
         {
             return redirectView( request, VIEW_MANAGE_FORMS );
         }
+        checkUserPermission( Form.RESOURCE_TYPE, String.valueOf( nId ), FormsResourceIdService.PERMISSION_MODIFY_PARAMS, request);
+        
 
         Form formToBeModified = FormHome.findByPrimaryKey( nId );
 
@@ -490,9 +511,11 @@ public class FormJspBean extends AbstractJspBean
      * @param request
      *            The Http request
      * @return The Jsp URL of the process result
+     * @throws fr.paris.lutece.portal.service.admin.AccessDeniedException
+     *            AccessDeniedException if the user isn'y authorized to process the modification of params of a form
      */
     @Action( ACTION_MODIFY_FORM )
-    public String doModifyForm( HttpServletRequest request )
+    public String doModifyForm( HttpServletRequest request ) throws AccessDeniedException
     {
         int nId = -1;
         try
@@ -505,6 +528,8 @@ public class FormJspBean extends AbstractJspBean
             addError( ERROR_FORM_NOT_UPDATED, getLocale( ) );
             return redirectView( request, VIEW_MANAGE_FORMS );
         }
+        
+        checkUserPermission( Form.RESOURCE_TYPE, String.valueOf( nId ), FormsResourceIdService.PERMISSION_MODIFY_PARAMS, request);
 
         _form = FormHome.findByPrimaryKey( nId );
 
@@ -563,5 +588,5 @@ public class FormJspBean extends AbstractJspBean
         return bIsFormValid;
 
     }
-
+    
 }
